Rename 'metabase_analytics' to 'analytics'
