+++ resources/frontend_client/app/lib/analytics.js
